added reservable module and changed related logic in admin and blocks
